dom/indexedDB/test/test_disable_indexeddb.html

<!DOCTYPE html>
<html>
<head>
  <title>Indexed Database Disable Pref test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>

<body>
  <script>
   add_task(function* () {
       for (let enable of [false, true]) {
           yield SpecialPowers.pushPrefEnv(
               {'set': [['dom.indexedDB.enabled', enable]]});
           if (!enable) {
               is(indexedDB, null, "indexedDB should be null");
           } else {
               isnot(indexedDB, null, "indexedDB should be non-null");
           }
       }
   });
  </script>
</body>

</html>
